+++ src/shared/types/records.types.ts
@@ -4,6 +4,7 @@ import { Category } from "./categories.types";
 import { DetailedError } from "./global.types";
 import { object, string } from "yup";
 
+//#region General interfaces
 export type TypeOfRecord = 'expense' | 'income' | 'transfer';
 
 export type TransferRecord = {
@@ -14,9 +15,15 @@ export type TransferRecord = {
 export type IndebtedPeople = {
   _id?: string;
   name: string;
-  amount: string;
-  amountPaid: string;
+  amount: number;
+  amountPaid: number;
   isPaid: boolean;
+}
+
+export type IndebtedPeopleUI = IndebtedPeople & {
+  amountFormatted: string
+  amountPaidFormatted: string;
+  remainingAmountFormatted: string;
 }
 
 export type AccountRecord = {
@@ -63,6 +70,8 @@ export type BankMovement = AccountRecord & {
   expensesPaid?: ExpensePaid[];
 }
 
+//#region Paylaod and responses interfaces
+
 export type GetAccountPayload = {
   accountId: string;
   month: string;
@@ -87,7 +96,7 @@ export type CreateExpensePayload = {
   description: string;
   indebtedPeople: IndebtedPeople[];
   isPaid: boolean;
-  linkedBudgets: Budget[];
+  linkedBudgets: string[];
   shortName: string;
   subCategory: string;
   tag: string[];
@@ -118,9 +127,15 @@ export type GetCurrentMonthRecordsResponse = {
   records: BankMovement[];
 }
 
+//#region Data form - schemas
+
 export type CreateExpenseDataForm = {
   shortDescription: string
   description?: string | null | undefined
+}
+
+export type AddIndebtedPeopleDataForm = {
+  name: string
 }
 
 const shortNameValidation = string()
@@ -141,3 +156,10 @@ export const CreateExpenseSchema = object().shape({
 }, [
   ["description", "description"]
 ])
+
+export const AddIndebtedPeopleSchema = object().shape({
+  name: string()
+    .required('Por favor, ingrese el nombre de la persona')
+    .min(3, 'El nombre debe contener más de 3 caracteres')
+    .max(100, 'El nombre debe contener menos de 100 caracteres.'),
+})